refactor(swarm-builder): concentrate the storer cone behind one reserve seam
Extract the whole storer build path (reserve, puller, redistribution and
pullsync wiring, the cache-then-reserve serve view, the storer config and
builder, and the launch path) out of the shared launch, node, and config modules
into one gated `storer` module in vertex-swarm-builder. The shared launch path is
now capability-agnostic: `build_client_backed_node` is generic over a `NodeFlavour`
seam, so no per-statement storer cfg threads through it and the client builder
carries no storer-only fields or reserve seams. The storer plugs in through the
flavour plus a `StorerNodeBuilder` that wraps the client builder.

Rename the inner builder feature `storer-core` to `reserve` (the storer composite
feature, the node `storer` feature, and `SwarmNodeType::Storer` are unchanged).
This collapses the storer cfg sites in the builder from roughly seventy to four
re-export gates in lib.rs. Behaviour is preserved; the external feature surface
(default client, `--features storer`, vertex-ffi, wasm) is unchanged.
