+++ crates/ffi/AGENTS.md
@@ -16,7 +16,7 @@ Root-level rules in `/AGENTS.md` apply here too, plus `docs/agents/api-surface.m
 ## Dos
 
 - Keep the API thin. The crate is a boundary, not a place for logic. Build the client through the highest-level builder entry point (`vertex_swarm_builder::DefaultClientBuilder`); drive chunks through `SwarmChunkSender` and `SwarmChunkProvider`.
-- FFI is a crate, not a feature (see the Feature and cfg contract in `/AGENTS.md`): it never enables the builder's `storer-core`, so the storer cone stays out, which the cone guard asserts. It does pull `vertex-swarm-builder` and therefore `vertex-storage-redb`: the builder is the native full-stack entry point and the FFI client is native-only, so redb stays as the persistent-cache backend a native mobile embedder will want (the native analogue of the browser's IndexedDB cache). The lighter `vertex_swarm_node::ClientLauncher` path keeps the verifying, selector-aware chunk provider in the builder, so adopting it would mean moving those provider types first; not worth it for a redb trim.
+- FFI is a crate, not a feature (see the Feature and cfg contract in `/AGENTS.md`): it never enables the builder's `reserve`, so the storer cone stays out, which the cone guard asserts. It does pull `vertex-swarm-builder` and therefore `vertex-storage-redb`: the builder is the native full-stack entry point and the FFI client is native-only, so redb stays as the persistent-cache backend a native mobile embedder will want (the native analogue of the browser's IndexedDB cache). The lighter `vertex_swarm_node::ClientLauncher` path keeps the verifying, selector-aware chunk provider in the builder, so adopting it would mean moving those provider types first; not worth it for a redb trim.
 - Reconstruct strong types (`StampedChunk`, `ChunkAddress`, `Stamp`) immediately on entry. Raw bytes and strings live only in the `api::types` shapes; never let them flow into internal logic.
 - Generate the C ABI from the Rust `api` module via flutter_rust_bridge. There is no hand-maintained parallel C header.
 - Annotate exported items with `#[frb(...)]` so codegen sees the right shape. `#[frb(opaque)]` for handles, `#[frb(non_opaque)]` for plain data, `#[frb(ignore)]` on private helper structs in `src/api` so the codegen does not generate glue for them.
