+++ crates/swarm/builder/src/storer/composite.rs
